+++ prowler/providers/aws/services/wafv2/wafv2_service.py
@@ -25,26 +25,27 @@ class WAFv2(AWSService):
 
     def _list_web_acls_global(self):
         logger.info("WAFv2 - Listing Global Web ACLs...")
-        try:
-            regional_client = self.regional_clients["us-east-1"]
-            for wafv2 in regional_client.list_web_acls(Scope="CLOUDFRONT")["WebACLs"]:
-                if not self.audit_resources or (
-                    is_resource_filtered(wafv2["ARN"], self.audit_resources)
-                ):
-                    arn = wafv2["ARN"]
-                    self.web_acls[arn] = WebAclv2(
-                        arn=arn,
-                        name=wafv2["Name"],
-                        id=wafv2["Id"],
-                        albs=[],
-                        user_pools=[],
-                        scope=Scope.CLOUDFRONT,
-                        region="us-east-1",
-                    )
-        except Exception as error:
-            logger.error(
-                f"us-east-1 -- {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
-            )
+        if "us-east-1" in self.regional_clients:
+            try:
+                    regional_client = self.regional_clients["us-east-1"]
+                    for wafv2 in regional_client.list_web_acls(Scope="CLOUDFRONT")["WebACLs"]:
+                        if not self.audit_resources or (
+                            is_resource_filtered(wafv2["ARN"], self.audit_resources)
+                        ):
+                            arn = wafv2["ARN"]
+                            self.web_acls[arn] = WebAclv2(
+                                arn=arn,
+                                name=wafv2["Name"],
+                                id=wafv2["Id"],
+                                albs=[],
+                                user_pools=[],
+                                scope=Scope.CLOUDFRONT,
+                                region="us-east-1",
+                            )
+            except Exception as error:
+                logger.error(
+                    f"us-east-1 -- {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
+                )
 
     def _list_web_acls_regional(self, regional_client):
         logger.info("WAFv2 - Listing Regional Web ACLs...")
@@ -130,23 +131,18 @@ class WAFv2(AWSService):
             try:
                 rules = get_web_acl.get("WebACL", {}).get("Rules", [])
                 for rule in rules:
-                    name = rule.get("Name", "")
-                    metrics_enabled = rule.get("VisibilityConfig", {}).get(
+                    new_rule = Rule(
+                        name=rule.get("Name", ""), cloudwatch_metrics_enabled=rule.get("VisibilityConfig", {}).get(
                         "CloudWatchMetricsEnabled", False
                     )
-                    new_rule = Rule(
-                        name=name, cloudwatch_metrics_enabled=metrics_enabled
                     )
-                    if (
-                        "RuleGroupReferenceStatement" in rule["Statement"]
-                        and "ARN" in rule["Statement"]["RuleGroupReferenceStatement"]
-                    ):
+                    if rule.get("Statement",{}).get("RuleGroupReferenceStatement",{}).get("ARN"):
                         acl.rule_groups.append(new_rule)
                     else:
                         acl.rules.append(new_rule)
 
             except Exception as error:
-                logger.warning(
+                logger.error(
                     f"{acl.region} -- {error.__class__.__name__}[{error.__traceback__.tb_lineno}]: {error}"
                 )
